requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

opencv_python == 4.5.5.64
pafy == 0.5.5
python_vlc == 3.0.16120
